vision-transformer/requirements.txt
tensorflow-datasets
tensorflow-addons
sklearn
